Friends are now displayed
In the main window of the application there is now a list of friends that are received from the server. But for now, only those friends with whom you are mutually friends are displayed. If you have sent a friend application, or a friend request to you, they are not displayed. I'm working on this.